% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_colors.R
\name{get_theme_colors}
\alias{get_theme_colors}
\title{Extract colors from bslib theme}
\usage{
get_theme_colors(theme = my_theme())
}
\arguments{
\item{theme}{A bslib theme object (defaults to my_theme())}
}
\value{
Named list of color values
}
\description{
Extract colors from bslib theme
}
