% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fonts.R
\name{get_theme_fonts}
\alias{get_theme_fonts}
\title{Extract fonts from bslib theme}
\usage{
get_theme_fonts(theme)
}
\arguments{
\item{theme}{A bslib theme object (defaults to my_theme())}
}
\value{
Named list of font values
}
\description{
Extract fonts from bslib theme
}
