% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{luwi_ggplotly}
\alias{luwi_ggplotly}
\title{Custom ggplotly function matching bslib theme}
\usage{
luwi_ggplotly(p, theme = my_theme(), base_size = 14, tooltip = "y")
}
\arguments{
\item{p}{a ggplot object}

\item{theme}{A bslib theme object (defaults to my_theme())}

\item{base_size}{Base font size}

\item{tooltip}{which variable and corresponding value to display on hover}
}
\value{
A ggplot2 theme object
}
\description{
Custom ggplotly function matching bslib theme
}
