% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{my_simple_page}
\alias{my_simple_page}
\title{Simple page without sidebar}
\usage{
my_simple_page(title, ...)
}
\arguments{
\item{title}{Application title}

\item{...}{Page content (UI elements)}
}
\value{
A Shiny UI object
}
\description{
Creates a full-width page with your branded theme.
}
