% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{scale_color_luwi_c}
\alias{scale_color_luwi_c}
\title{ggplot2 color scale for continuous data}
\usage{
scale_color_luwi_c(type = "warm", ...)
}
\arguments{
\item{type}{Type of palette: "warm", "cool", or "green"}

\item{...}{Additional arguments passed to scale_color_gradientn}
}
\description{
ggplot2 color scale for continuous data
}
