% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{scale_color_luwi_d}
\alias{scale_color_luwi_d}
\title{ggplot2 color scale for discrete/categorical data}
\usage{
scale_color_luwi_d(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_manual}
}
\description{
ggplot2 color scale for discrete/categorical data
}
