% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_colors.R
\name{scale_color_luwi_discrete}
\alias{scale_color_luwi_discrete}
\title{Get discrete color palette for categorical data}
\usage{
scale_color_luwi_discrete(theme = my_theme(), n = NULL)
}
\arguments{
\item{theme}{A bslib theme object (defaults to my_theme())}

\item{n}{Number of colors needed (max 15)}
}
\value{
Character vector of hex colors
}
\description{
Get discrete color palette for categorical data
}
