% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{scale_color_luwi_div}
\alias{scale_color_luwi_div}
\title{ggplot2 color scale for diverging data (hot/cold)}
\usage{
scale_color_luwi_div(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_gradientn}
}
\description{
ggplot2 color scale for diverging data (hot/cold)
}
