% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_colors.R
\name{scale_color_luwi_diverging}
\alias{scale_color_luwi_diverging}
\title{Get diverging color palette (e.g., for hot/cold data)}
\usage{
scale_color_luwi_diverging(theme = my_theme(), n = 11, reverse = FALSE)
}
\arguments{
\item{theme}{A bslib theme object (defaults to my_theme())}

\item{n}{Number of colors to generate (default: 11, should be odd)}

\item{reverse}{Reverse the palette direction}
}
\value{
Character vector of hex colors
}
\description{
Get diverging color palette (e.g., for hot/cold data)
}
