% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{scale_fill_luwi_d}
\alias{scale_fill_luwi_d}
\title{ggplot2 fill scale for discrete/categorical data}
\usage{
scale_fill_luwi_d(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_fill_manual}
}
\description{
ggplot2 fill scale for discrete/categorical data
}
