% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{scale_fill_luwi_div}
\alias{scale_fill_luwi_div}
\title{ggplot2 fill scale for diverging data (hot/cold)}
\usage{
scale_fill_luwi_div(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_fill_gradientn}
}
\description{
ggplot2 fill scale for diverging data (hot/cold)
}
