% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{theme_luwi}
\alias{theme_luwi}
\title{Custom ggplot2 theme matching bslib theme}
\usage{
theme_luwi(theme = my_theme(), base_size = 14)
}
\arguments{
\item{theme}{A bslib theme object (defaults to my_theme())}

\item{base_size}{Base font size}
}
\value{
A ggplot2 theme object
}
\description{
Custom ggplot2 theme matching bslib theme
}
